requirements.txt
certifi==2018.8.24
chardet==3.0.4
click==6.7
coreapi==2.3.3
coreschema==0.0.4
dateutils==0.6.6
Django==2.*
django-autoslug==1.9.3
django-autoslug-iplweb==1.9.4.dev0
django-chroniker==1.0.10
django-cities-light==3.5.0
django-cors-headers==2.4.0
django-rest-multiple-models==2.1.1
django-rest-swagger==2.1.2
djangorestframework==3.9.4
docopt==0.6.2
drf-openapi==1.3.0
flake8==3.5.0
idna==2.7
itypes==1.1.0
jsonpointer==2.0
jsonschema==2.6.0
MarkupSafe==1.0
mccabe==0.6.1
mysqlclient==1.3.13
openapi-codec==1.3.2
pbr==4.2.0
Pillow==6.0.0
pipenv==2018.7.1
pipreqs==0.4.9
progressbar2==3.38.0
psutil==5.4.7
pycodestyle==2.4.0
pyflakes==1.6.0
python-dateutil==2.7.3
python-utils==2.3.0
pytz==2018.5
requests==2.22.0
simplejson==3.16.0
six==1.11.0
stevedore==1.29.0
toposort==1.5
Unidecode==1.0.22
uritemplate==3.0.0
urllib3==1.25.3
virtualenv==16.0.0
virtualenv-clone==0.3.0
virtualenvwrapper==4.8.2
virtualenvwrapper-win==1.2.5
yarg==0.1.9
